Impl Customer Book1 UI And Interaction
- BP_Customer Book1의 UI 설정 및 손님 UI 클릭 동작 구현 완료
- 손님 UI 클릭 시 BP_Customer Book2가 생성되며 클릭한 손님의 테이블 인덱스 번호가 전달되도록 구현

+++ Source/CookYourWay/Actor/Customer.cpp
@@ -169,7 +169,7 @@ bool ACustomer::GetDestByLoyalty()
 	FCustomerBistroKey PlayerBistroKey = CustomerDataManagerSystem->GetCustomerBistroKey(CustName, VillageManagerSystem->PlayerBistroAreaID);
 	if (CustomerDataManagerSystem->IsRegularCustMap.Find(PlayerBistroKey)) {
 		int32 Rand = UKismetMathLibrary::RandomIntegerInRange(1, 100);
-		float Loyalty = *CustomerDataManagerSystem->LoyaltyMap.Find(PlayerBistroKey);
+		int32 Loyalty = *CustomerDataManagerSystem->LoyaltyMap.Find(PlayerBistroKey);
 
 		if (Rand <= Loyalty) {
 			VisitDest = *VillageManager->AreaLocMap.Find(VillageManagerSystem->PlayerBistroAreaID);
@@ -182,7 +182,7 @@ bool ACustomer::GetDestByLoyalty()
 		FCustomerBistroKey CompetitorKey = CustomerDataManagerSystem->GetCustomerBistroKey(CustName, CompetitorData.AreaID);
 
 		if (CustomerDataManagerSystem->IsRegularCustMap.Find(CompetitorKey)) {
-			float Loyalty = *CustomerDataManagerSystem->LoyaltyMap.Find(CompetitorKey);
+			int32 Loyalty = *CustomerDataManagerSystem->LoyaltyMap.Find(CompetitorKey);
 
 			if (Rand <= Loyalty) {
 				VisitDest = *VillageManager->AreaLocMap.Find(CompetitorData.AreaID);
